# Conflicts:
#	CHANGELOG
#	pom.xml
#	src/org/zoodb/internal/DataDeSerializer.java
#	src/org/zoodb/internal/DataSerializer.java
#	src/org/zoodb/internal/server/ObjectReader.java
#	src/org/zoodb/jdo/impl/PersistenceManagerFactoryImpl.java
#	src/org/zoodb/jdo/impl/QueryImpl.java
#	src/org/zoodb/jdo/impl/TransactionImpl.java

+++ src/org/zoodb/profiling/api/impl/ProfilingQueryListener.java
@@ -26,7 +26,7 @@ public class ProfilingQueryListener implements IQueryListener {
 		//initialize a new one for each connection --> advantage: queries could run in paralled over multiple connections and IO is still correct.
 		dbStats = new DBStatistics((Session) query.getPersistenceManager().getDataStoreConnection().getNativeConnection());
 		int queryId = QueryManager.getId(query);
-		int pageCountBegin = dbStats.getStorageDataPageReadCount();
+		long pageCountBegin = dbStats.getStorageDataPageReadCount();
 		ProfilingManager.getInstance().getQueryManager().insertProfile(queryId, query,pageCountBegin);
 	}
 
@@ -34,7 +34,7 @@ public class ProfilingQueryListener implements IQueryListener {
 	public void afterExecution(QueryImpl query) {
 		dbStats = new DBStatistics((Session) query.getPersistenceManager().getDataStoreConnection().getNativeConnection());
 		int queryId = QueryManager.getId(query);
-		int pageCountEnd = dbStats.getStorageDataPageReadCount();
+		long pageCountEnd = dbStats.getStorageDataPageReadCount();
 		ProfilingManager.getInstance().getQueryManager().updateProfile(queryId, query,pageCountEnd);
 	}
